+++ jdbc/src/main/java/nextstep/jdbc/JdbcContext.java
@@ -1,0 +1,47 @@
+package nextstep.jdbc;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class JdbcContext {
+
+    private static final Logger log = LoggerFactory.getLogger(JdbcContext.class);
+
+    private final DataSource dataSource;
+
+    public JdbcContext(DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    public void workWithStatementStrategy(StatementStrategy statementStrategy) {
+        Connection conn = null;
+        PreparedStatement pstmt = null;
+        try {
+            conn = dataSource.getConnection();
+            pstmt = statementStrategy.makePreparedStatement(conn);
+
+            pstmt.executeUpdate();
+        } catch (SQLException e) {
+            log.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        } finally {
+            try {
+                if (pstmt != null) {
+                    pstmt.close();
+                }
+            } catch (SQLException ignored) {
+            }
+
+            try {
+                if (conn != null) {
+                    conn.close();
+                }
+            } catch (SQLException ignored) {
+            }
+        }
+    }
+}
