+++ jdbc/src/main/java/nextstep/jdbc/StatementStrategy.java
@@ -1,4 +1,4 @@
-package com.techcourse.dao;
+package nextstep.jdbc;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
